refactor: make `ArtworkListImage` reusable
