+++ src/app/Components/ArtworkLists/components/ArtworkListImage.tsx
